Test: Move client/lib/mixin/i18n tests to single test runner
