--- client/lib/mixins/i18n/Makefile
@@ -1,12 +1,0 @@
-NODE_BIN := $(shell npm bin)
-MOCHA ?= $(NODE_BIN)/mocha
-BASE_DIR := $(NODE_BIN)/../..
-NODE_PATH := test:$(BASE_DIR)/client
-COMPILERS ?= jsx:babel/register
-REPORTER ?= spec
-UI ?= bdd
-
-test:
-	@NODE_ENV=test NODE_PATH=$(NODE_PATH) $(MOCHA) --compilers $(COMPILERS) --reporter $(REPORTER) --ui $(UI)
-
-.PHONY: test
